floats

+++ src/ft_pow.c
@@ -10,9 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_pow(int a, int b)
+long long int	ft_pow(int a, int b)
 {
-	int res;
+	long long int res;
 
 	res = 1;
 	if (b == 0)
